feat: Add likes_count column to threads and posts table
This commit adds the `likes_count` column to the `threads` and `posts` tables in the database. The column is of type `integer` and has a default value of 0. This change allows for tracking the number of likes received by threads and posts, enhancing the engagement and interaction within the application.